+++ C# Fundamentals/Topics/OOPM/Banking Application/SavingsAccount.cs
@@ -13,7 +13,7 @@ namespace C__Fundamentals.Topics.OOPM.Banking_Application
         {
         }
 
-        public override void Withdraw(int amount)
+        public override void Withdraw(double amount)
         {
             if(balance-amount > minBalance && amount > 0)
             {
